Dev (#2)
* Add Jekyll deployment badge to README

Inserted a GitHub Actions badge to display the build status for Jekyll deployment with GitHub Pages dependencies preinstalled. This provides visibility into the deployment process directly from the README.md file, ensuring users and contributors can easily monitor the action's status.

* Add initial README for Aspire Blog Application

This README provides an overview of the Aspire Blog Application, detailing its main features such as blog post management and integration with MongoDB using Docker for testing. It includes badges highlighting project status, build, and test processes, along with guidelines for contributing and engaging with the project.

* Remove CODE_METRICS.md file

The CODE_METRICS.md file has been deleted as it was automatically generated and should not be edited manually. By removing this file, we ensure that unnecessary maintenance or potential confusion is avoided. This change helps streamline the project management process.

* Update paths for contributing and code of conduct links

Modified the file paths for the contributing guide and code of conduct in the README to reflect their correct locations under the 'doc' directory. This ensures that users can navigate to and access these documents without encountering broken links.