Remove redundant workflows and branch triggers (#20)
* "Remove 'dev' branch from CodeQL workflow triggers"

The 'dev' branch was removed from the workflow triggers and analysis paths in the CodeQL configuration file. This change streamlines the workflow to focus on the 'main' branch for builds and analysis.

* Update branch trigger in code-metrics workflow

Changed the workflow trigger branch from 'main' to 'dev' to align with the updated development process. This ensures the code-metrics workflow runs only on the 'dev' branch.

* Remove redundant GitHub workflows for code metrics and CodeQL.

Deleted `.github/workflows/code-metrics.yml` and `.github/workflows/codeql-analysis.yml` as they are no longer needed. This simplifies the repository and eliminates unused automation scripts.